fix: refactor code, attach expres async router, fix typescript issues
